move parameter builder to sharding and encrypt packages
