+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/encrypt/parameter/EncryptParameterBuilderFactory.java
@@ -15,14 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.parameter.rewriter.encrypt;
+package org.apache.shardingsphere.core.rewrite.encrypt.parameter;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.core.metadata.table.TableMetas;
 import org.apache.shardingsphere.core.optimize.statement.SQLStatementContext;
-import org.apache.shardingsphere.core.rewrite.parameter.builder.ParameterBuilder;
-import org.apache.shardingsphere.core.rewrite.parameter.rewriter.ParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.encrypt.parameter.impl.EncryptAssignmentParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.encrypt.parameter.impl.EncryptInsertValueParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.encrypt.parameter.impl.EncryptPredicateParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.parameter.ParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.ParameterRewriter;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.EncryptRuleAware;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.QueryWithCipherColumnAware;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.TableMetasAware;
