+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/parameter/ParameterBuilder.java
@@ -15,26 +15,31 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.parameter.rewriter;
+package org.apache.shardingsphere.core.rewrite.parameter;
 
-import org.apache.shardingsphere.core.optimize.statement.SQLStatementContext;
-import org.apache.shardingsphere.core.rewrite.parameter.builder.ParameterBuilder;
+import org.apache.shardingsphere.core.route.type.RoutingUnit;
 
 import java.util.List;
 
 /**
- * Parameter rewriter.
+ * Parameter builder.
  *
- * @author zhangliang
+ * @author panjuan
  */
-public interface ParameterRewriter {
+public interface ParameterBuilder {
     
     /**
-     * Rewrite SQL parameters.
-     * 
-     * @param sqlStatementContext SQL statement context
-     * @param parameters SQL parameters
-     * @param parameterBuilder parameter builder
+     * Get parameters.
+     *
+     * @return parameters
      */
-    void rewrite(SQLStatementContext sqlStatementContext, List<Object> parameters, ParameterBuilder parameterBuilder);
+    List<Object> getParameters();
+    
+    /**
+     * Get parameters.
+     *
+     * @param routingUnit routing unit
+     * @return parameters
+     */
+    List<Object> getParameters(RoutingUnit routingUnit);
 }
