+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/parameter/impl/StandardParameterBuilder.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.parameter.builder.impl;
+package org.apache.shardingsphere.core.rewrite.parameter.impl;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.core.rewrite.parameter.builder.ParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.ParameterBuilder;
 import org.apache.shardingsphere.core.route.type.RoutingUnit;
 
 import java.util.LinkedHashMap;
