+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/sharding/parameter/ShardingParameterBuilderFactory.java
@@ -15,14 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rewrite.parameter.rewriter.sharding;
+package org.apache.shardingsphere.core.rewrite.sharding.parameter;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.core.metadata.table.TableMetas;
-import org.apache.shardingsphere.core.rewrite.parameter.builder.ParameterBuilder;
-import org.apache.shardingsphere.core.rewrite.parameter.builder.impl.GroupedParameterBuilder;
-import org.apache.shardingsphere.core.rewrite.parameter.rewriter.ParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.parameter.ParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.impl.GroupedParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.ParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.sharding.parameter.impl.GeneratedKeyInsertValueParameterRewriter;
+import org.apache.shardingsphere.core.rewrite.sharding.parameter.impl.PaginationParameterRewriter;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.SQLRouteResultAware;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.ShardingRuleAware;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.TableMetasAware;
