+++ sharding-core/sharding-core-rewrite/src/test/java/org/apache/shardingsphere/core/rewrite/parameter/builder/group/GroupedParameterBuilderTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.core.rewrite.parameter.builder.group;
 
-import org.apache.shardingsphere.core.rewrite.parameter.builder.impl.GroupedParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.impl.GroupedParameterBuilder;
 import org.apache.shardingsphere.core.route.router.sharding.condition.ShardingCondition;
 import org.apache.shardingsphere.core.route.router.sharding.condition.ShardingConditions;
 import org.apache.shardingsphere.core.route.type.RoutingUnit;
