+++ sharding-core/sharding-core-rewrite/src/test/java/org/apache/shardingsphere/core/rewrite/parameter/builder/standard/StandardParameterBuilderTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.core.rewrite.parameter.builder.standard;
 
-import org.apache.shardingsphere.core.rewrite.parameter.builder.impl.StandardParameterBuilder;
+import org.apache.shardingsphere.core.rewrite.parameter.impl.StandardParameterBuilder;
 import org.junit.Before;
 import org.junit.Test;
